% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_responses_table.R
\name{mod_responses_table_ui}
\alias{mod_responses_table_ui}
\alias{mod_responses_table_server}
\title{mod_responses_table_ui and mod_responses_table_server}
\usage{
mod_responses_table_ui(id)

mod_responses_table_server(input, output, session, rv, global, res_auth)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
